Moved types aliases to a different file.
