+++ src/api/config/api-query-config.ts
@@ -1,4 +1,4 @@
-import { QueryResult, TMQuery, TQuery } from '../../persistence/primary/query/ant-primary-query-manager';
+import { QueryResult, TMQuery, TQuery } from '../../persistence/primary/query/query-types';
 import { Entity } from '../../model/entity';
 
 export interface ApiQueryConfig<TEntity extends Entity, TQueryResult extends QueryResult> {
